Add video media type to twitter
The variants from twitter contain a bunch of URLs to either MP4 videos
or an mu3 video playlist. To find the largest video, I parse the URLs
which contain dimension info like "720x720". I parse the numbers and
multiply them together to get the video size. Then that number is used
to find the largest variant. If the variant doesn't have the dimension
string in it, this assumes it's not a video (and probably an m3u
playlist) and returns 0 for the size.